Remove wrong error detection for output on stderr
It is possible that a the quartus_sta command produces some output
on stderr even if there was no (serious) error. But the tcl exec
command in the pin-assignment script will treat any such output
as an error.

As the script is automatically generated and as such cannot be
patched, run the script with the correct interpreter to prevent
starting it from within tcl.